Centralized deprecated code
