+++ Sources/InterposeKit/Deprecated/Hook+Deprecated.swift
